Move coordinate/grid functions out of iharm model. iharm and all analytic models compile.
