+++ model/sphere/model.h
@@ -22,7 +22,6 @@
 #define B1       5
 #define B2       6
 #define B3       7
-#define NPRIM    8
 
 #define SMALL (1.e-40)
 #define MMW   (0.5)
